+++ delphi-frontend/src/main/java/au/com/integradev/delphi/msbuild/DelphiProject.java
@@ -38,5 +38,9 @@ public interface DelphiProject {
 
   List<Path> getDebugSourceDirectories();
 
+  List<Path> getLibraryPathDirectories();
+
+  List<Path> getBrowsingPathDirectories();
+
   Map<String, String> getUnitAliases();
 }