+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.13.0] - 2025-02-05
+
 ### Added
 
 - **API:** `AnonymousMethodNode::getDeclarationSection` method.
@@ -1442,7 +1444,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - False-positives in 13 rules.
 - Logic errors in existing "semantic analysis" that caused scan failures.
 
-[Unreleased]: https://github.com/integrated-application-development/sonar-delphi/compare/v1.12.2...HEAD
+[Unreleased]: https://github.com/integrated-application-development/sonar-delphi/compare/v1.13.0...HEAD
+[1.13.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v1.12.2...v1.13.0
 [1.12.2]: https://github.com/integrated-application-development/sonar-delphi/compare/v1.12.1...v1.12.2
 [1.12.1]: https://github.com/integrated-application-development/sonar-delphi/compare/v1.12.0...v1.12.1
 [1.12.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v1.11.0...v1.12.0